Move serde helpers into specific class definitions
